backend/session-service-old/requirements.txt
aiohttp>=3.8.5
aiohttp-cors>=0.7.0
asyncio>=3.4.3
PyJWT>=2.7.0
pydantic==2.5.0
kubernetes==29.0.0
pyyaml==6.0.1
grpcio>=1.62.0
grpcio-tools==1.62.0
protobuf>=4.21.0
asyncpg>=0.27.0
uvloop==0.19.0
deprecated>=1.2.0
# PROMETHEUS TRACING
prometheus-client>=0.16.0
# JAEGER TRACING
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-instrumentation-aiohttp-client>=0.40b0
opentelemetry-instrumentation-asyncpg>=0.40b0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
